requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.36.0
motor==3.7.1
pymongo==4.15.1
httpx==0.28.1
python-multipart==0.0.20
pydantic==2.10.3
python-dotenv==1.1.0
google-generativeai==0.3.2
